feat(backup-script): commit msg passing via args
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh
	modified:   i3/.config/i3/config
	modified:   neovim/.config/nvim/lua/user/gitsigns.lua
	modified:   scripts/.config/scripts/toggle_monitor.sh

no changes added to commit (use "git add" and/or "git commit -a")

Regular Backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

Regular Backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

Regular Backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is ahead of 'origin/master' by 1 commit.
  (use "git push" to publish your local commits)

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is ahead of 'origin/master' by 2 commits.
  (use "git push" to publish your local commits)

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

On branch master
Your branch is ahead of 'origin/master' by 3 commits.
  (use "git push" to publish your local commits)

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")

updated backup.sh script for args

On branch master
Your branch is ahead of 'origin/master' by 4 commits.
  (use "git push" to publish your local commits)

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup.sh

no changes added to commit (use "git add" and/or "git commit -a")