cron backup
On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   crontabs
	modified:   neovim/.config/nvim/lua/user/plugins/nvim-tree.lua

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	cat.txt

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add/rm <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	deleted:    cat.txt

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   backup

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   neovim/.config/nvim/lua/user/plugins/nvim-tree.lua

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	neovim/.config/nvim/file

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   crontabs

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   neovim/.config/nvim/lua/user/plugins/plugins.lua

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   neovim/.config/nvim/lua/user/plugins/plugins.lua

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   neovim/.config/nvim/lua/user/plugins/plugins.lua

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   neovim/.config/nvim/lua/user/plugins/plugins.lua

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   scripts/.config/scripts/lock.sh

no changes added to commit (use "git add" and/or "git commit -a")