fuckup
On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   scripts/.config/scripts/eye_break_notifier

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	.stow-local-ignore
	Xresources/.config/
	Xresources/.fzf
	Xresources/.fzf.bash
	Xresources/.fzf.zsh
	Xresources/.oh-my-zsh
	Xresources/.p10k.zsh
	Xresources/.tmux.conf
	Xresources/.zsh_history
	Xresources/.zshenv
	Xresources/.zshrc
	Xresources/powerlevel10k

no changes added to commit (use "git add" and/or "git commit -a")

fucked up my p10k config

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add/rm <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	deleted:    .stow-local-ignore
	modified:   neovim/.config/nvim/lua/user/plugins/init.lua
	modified:   neovim/.config/nvim/lua/user/plugins/nvim-tree.lua
	modified:   neovim/.config/nvim/lua/user/plugins/plugins.lua
	modified:   zsh/.zshrc

no changes added to commit (use "git add" and/or "git commit -a")

Regular backup

On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   neovim/.config/nvim/lua/user/plugins/lazygit_toggle_term.lua

no changes added to commit (use "git add" and/or "git commit -a")